+++ changelog.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 The versioning pattern is `major.minor.patch`.
 
 ---
+## v0.1.4
+- fix issue with mounts in add_task method.
+
+## v0.1.3
+- added unit tests in `tests` folder.
+
 ## v0.1.2
 - added download_task_output parameter to `CloudClient.monitor_job()` to download stdout and stderr of each task when the task completes.